services/communication-service/requirements.txt
fastapi
uvicorn[standard]
prometheus-client
prometheus-fastapi-instrumentator
aio-pika
motor
pymongo
pyjwt
pydantic
